Moved duckdb connection
