+++ opencontext_py/apps/all_items/isamples/parquet.py
@@ -7,7 +7,7 @@ from duckdb.typing import *
 
 from django.conf import settings
 
-from opencontext_py.apps.all_items.isamples import duckdb_con
+from opencontext_py.libs import duckdb_con
 
 
 OC_RAW_TABLES = [
